requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
starlette>=0.45.3
python-dotenv>=1.0.1
fuzzywuzzy==0.18.0